add es8388 driver(unstable)
this is used to help you develop es8388 on vocore2 ultimate.
it is not fully completed, still need improve and tutorial to use.
for experience developer, this has enough information to debug and play with the es8388 codec.
but this driver is modified from es8328 driver in linux kernel, some function such as mixer can not work properly, it might not a bug, because I am a noob about alsa. :) hope later I can improve it.